requirements.txt
mysql-connector-python==8.0.29
pathlib2==2.3.5
python-dotenv==0.18.0
